+++ controllers/BillPayment/delete.js
@@ -2,9 +2,13 @@ const { ActivityUser } = require("../../helper/Activity.interface");
 const { DeleteBillPayment } = require("../../helper/BillPayment");
 const httpStatus = require("../../helper/Httplib");
 const { errorResponse, successResponse } = require("../../helper/Response");
+const { CheckPermissionDelete } = require("../../middlewares/permission");
 
 const deleteBillPayment = async (req, res) => {
   try {
+    if (CheckPermissionDelete(req, res, ActivityUser.BillPayment) === false) {
+      return errorResponse(res, httpStatus.unauthorized, "Unauthorized User");
+    }
     const { idBillPayment } = req.params;
 
     await DeleteBillPayment(req, idBillPayment);
