+++ controllers/DeliveryOrder/delete.js
@@ -1,11 +1,15 @@
 const { ActivityUser } = require("../../helper/Activity.interface");
 const { DeleteDeliveryOrder } = require("../../helper/DeliveryOrder/index");
 const httpStatus = require("../../helper/Httplib");
-const { successResponse } = require("../../helper/Response");
+const { successResponse, errorResponse } = require("../../helper/Response");
 const { CreateActivityUser } = require("../../helper/UserActivity");
+const { CheckPermissionDelete } = require("../../middlewares/permission");
 
 const deleteDeliveryOrder = async (req, res) => {
   try {
+    if (CheckPermissionDelete(req, res, ActivityUser.DeliveryOrder) === false) {
+      return errorResponse(res, httpStatus.unauthorized, "Unauthorized User");
+    }
     const { idDo } = req.params;
 
     await DeleteDeliveryOrder(req, idDo);
