+++ controllers/PurchaseOrder/delete.js
@@ -3,9 +3,13 @@ const httpStatus = require("../../helper/Httplib");
 const { deletePurchaseOrder } = require("../../helper/PurchaseOrder");
 const { errorResponse, successResponse } = require("../../helper/Response");
 const { CreateActivityUser } = require("../../helper/UserActivity");
+const { CheckPermissionDelete } = require("../../middlewares/permission");
 
 const deletePo = async (req, res) => {
   try {
+    if (CheckPermissionDelete(req, res, ActivityUser.PurchaseOrder) === false) {
+      return errorResponse(res, httpStatus.unauthorized, "Unauthorized User");
+    }
     const { idPo } = req.params;
 
     await deletePurchaseOrder(req, res, idPo);
