+++ controllers/SalesOrder/delete.js
@@ -3,9 +3,13 @@ const httpStatus = require("../../helper/Httplib");
 const { errorResponse, successResponse } = require("../../helper/Response");
 const { DeleteSalesOrder } = require("../../helper/SalesOrder");
 const { CreateActivityUser } = require("../../helper/UserActivity");
+const { CheckPermissionDelete } = require("../../middlewares/permission");
 
 const deleteSalesOrder = async (req, res) => {
   try {
+    if (CheckPermissionDelete(req, res, ActivityUser.SalesOrder) === false) {
+      return errorResponse(res, httpStatus.unauthorized, "Unauthorized User");
+    }
     const { idSo } = req.params;
 
     await DeleteSalesOrder(res, idSo);
